Pisp-Backend issues resolved

+++ src/main/java/com/wso2/finance/open/banking/pisp/impl/UserApiServiceImpl.java
@@ -18,6 +18,7 @@ import com.wso2.finance.open.banking.pisp.utilities.constants.Constants;
 import com.wso2.finance.open.banking.pisp.utilities.constants.ErrorMessages;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import javax.ws.rs.core.Response;
@@ -29,17 +30,10 @@ public class UserApiServiceImpl extends UserApiService {
 
     private Log log = LogFactory.getLog(UserApiServiceImpl.class);
 
-    /*
-    ===============================
-    This section handles the E-Shop
-    ===============================
-    */
-
     @Override
     public Response addNewEShop(EShopProfileDTO body) {
 
         EShop newEShop = UserMapping.createEshopInstance(body);
-
         UserManagementService eShopManagementService = new UserManagementService();
         InternalResponse eShopValidationResult = eShopManagementService.validateUsername(newEShop.getUsername(), Constants.E_SHOP);
         if (eShopValidationResult.isOperationSuccessful()) {
@@ -65,19 +59,16 @@ public class UserApiServiceImpl extends UserApiService {
     public Response eShopLogin(HttpServletRequest request, LoginCredentialsDTO body) {
 
         try {
-
             HttpSession session = request.getSession(true);
-            log.info("session is new :" + session.isNew());
-            log.info("session id" + session.getId());
-
             UserManagementService userManagementService = new UserManagementService();
 
             InternalResponse eshopLoginResponse = userManagementService.loginUser(body.getUsername(), body.getPassword(), Constants.E_SHOP);
             if (eshopLoginResponse.isOperationSuccessful()) {
                 Object sessionToken = SessionManager.generateSessionTokenForEShop(body.getUsername());
-                log.info("session token generated for E-shop" + sessionToken);
+                if (log.isDebugEnabled()) {
+                    log.debug("session token generated for E-shop" + body.getUsername());
+                }
                 session.setAttribute(Constants.SESSION_ID, sessionToken);
-
                 return Response.ok().entity(new ApiResponseMessage(ApiResponseMessage.OK, eshopLoginResponse.getMessage())).build();
             } else {
                 if (eshopLoginResponse.getMessage().equals(ErrorMessages.USERNAME_DOESNT_EXIST)) {
@@ -95,19 +86,15 @@ public class UserApiServiceImpl extends UserApiService {
     }
 
     @Override
-    public Response getEShopProfile(String username, String cookie, HttpServletRequest request) {
+    public Response getEShopProfile(String username, HttpServletRequest request) {
 
         HttpSession session = request.getSession(true);
-        log.info("session is new :" + session.isNew());
-        log.info("session id" + session.getId());
         String sessionToken = (String) session.getAttribute(Constants.SESSION_ID);
-        log.info("session Token " + sessionToken);
 
         InternalResponse response = SessionManager.validateSessionTokenOfEShop(username, sessionToken);
         if (response.isOperationSuccessful()) {
             UserManagementService userManagementService = new UserManagementService();
             EShop e_shop = userManagementService.getEshopUserProfle(username);
-            log.info("e_shop details merchant cat code" + e_shop.getMerchant().getMerchantCategoryCode());
             EShopProfileDTO eShopProfileDTO = UserMapping.getEShopProfileDTO(e_shop);
             return Response.ok().header(Constants.CONTENT_TYPE_HEADER, Constants.CONTENT_TYPE)
                     .entity(eShopProfileDTO).build();
@@ -119,22 +106,17 @@ public class UserApiServiceImpl extends UserApiService {
     }
 
     @Override
-    public Response updateEShopProfile(String username, HttpServletRequest request, String cookie, EShopProfileDTO body) {
+    public Response updateEShopProfile(String username, HttpServletRequest request, EShopProfileDTO body) {
 
         HttpSession session = request.getSession(true);
-        log.info("session is new :" + session.isNew());
-        log.info("session id" + session.getId());
         String sessionToken = (String) session.getAttribute(Constants.SESSION_ID);
-        log.info("session Token " + sessionToken);
 
         InternalResponse response = SessionManager.validateSessionTokenOfEShop(username, sessionToken);
         if (response.isOperationSuccessful()) {
             EShop updated_e_shop = UserMapping.createEshopInstance(body);
             UserManagementService userManagementService = new UserManagementService();
             userManagementService.updateEshopUserProfile(username, updated_e_shop);
-
-            return Response.ok().header(Constants.CONTENT_TYPE_HEADER, Constants.CONTENT_TYPE)
-                    .entity("E-shop user updated").build();
+            return Response.ok().entity(new ApiResponseMessage(ApiResponseMessage.OK, Constants.UPDATED_SUCCESSFULLY)).build();
         } else {
             return Response.serverError()
                     .entity(new ApiResponseMessage(ApiResponseMessage.ERROR, response.getMessage())).build();
@@ -142,34 +124,35 @@ public class UserApiServiceImpl extends UserApiService {
     }
 
     @Override
-    public Response deleteEShop(String username, String cookie) {
+    public Response deleteEShop(String username, HttpServletRequest request) {
 
-        UserManagementService userManagementService = new UserManagementService();
-        if (userManagementService.removeEshop(username)) {
-            return Response.ok().entity(new ApiResponseMessage(ApiResponseMessage.OK, Constants.ESHOP_DELETED)).build();
+        HttpSession session = request.getSession(true);
+        String sessionToken = (String) session.getAttribute(Constants.SESSION_ID);
+
+        InternalResponse response = SessionManager.validateSessionTokenOfEShop(username, sessionToken);
+        if (response.isOperationSuccessful()) {
+            UserManagementService userManagementService = new UserManagementService();
+            if (userManagementService.removeEshop(username)) {
+                return Response.ok().entity(new ApiResponseMessage(ApiResponseMessage.OK, Constants.ESHOP_DELETED)).build();
+            } else {
+                return Response.status(403).type("text/plain").entity(ErrorMessages.USERNAME_DOESNT_EXIST).build();
+            }
         } else {
-            return Response.status(403).type("text/plain").entity(ErrorMessages.USERNAME_DOESNT_EXIST).build();
+            return Response.serverError()
+                    .entity(new ApiResponseMessage(ApiResponseMessage.ERROR, response.getMessage())).build();
         }
-    }
 
-    /*
-    ============================
-    This section handles the PSU
-    ============================
-    */
+    }
 
     @Override
     public Response addNewPSU(PSUProfileDTO body) {
 
         PSU psu = UserMapping.createPSUInstance(body);
-
         UserManagementService psuManagementService = new UserManagementService();
         InternalResponse psuValidationResult = psuManagementService.validateUsername(psu.getUsername(), Constants.PSU);
 
         if (psuValidationResult.isOperationSuccessful()) {
-
             InternalResponse registration_result = psuManagementService.registerNewPSU(psu);
-
             if (registration_result.isOperationSuccessful()) {
                 return Response.ok()
                         .entity(new ApiResponseMessage(ApiResponseMessage.OK, registration_result.getMessage())).build();
@@ -177,7 +160,6 @@ public class UserApiServiceImpl extends UserApiService {
                 return Response.serverError()
                         .entity(new ApiResponseMessage(ApiResponseMessage.ERROR, registration_result.getMessage())).build();
             }
-
         } else {
             return Response.status(403).type("text/plain").entity(psuValidationResult.getMessage()).build();
         }
@@ -187,20 +169,17 @@ public class UserApiServiceImpl extends UserApiService {
     public Response loginPSU(String paymentInitReqId, HttpServletRequest request, LoginCredentialsDTO body) {
 
         try {
-
             HttpSession session = request.getSession(true);
-            log.info("session is new :" + session.isNew());
-            log.info("session id" + session.getId());
-
             UserManagementService userManagementService = new UserManagementService();
 
             InternalResponse psuLoginResponse = userManagementService.loginUser(body.getUsername(), body.getPassword(), Constants.PSU);
 
             if (psuLoginResponse.isOperationSuccessful()) {
                 Object sessionToken = SessionManager.generateSessionTokenForPSU(body.getUsername(), paymentInitReqId);
-                log.info("session token generated for PSU" + sessionToken);
+                if (log.isDebugEnabled()) {
+                    log.debug("session token generated for PSU" + body.getUsername());
+                }
                 session.setAttribute(Constants.SESSION_ID, sessionToken);
-
                 PaymentManagementService paymentManagementService = new PaymentManagementService();
                 paymentManagementService.updatePaymentWithPSU(paymentInitReqId, body.getUsername());
